+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Don't show page reload alert if attempt to load conversation fails.
+- Fixed custom emoji size in emoji reaction notification.
 
 ## [5.4.0] - 2026-05-27